+++ app/lib/screens/settings/settings_screen.dart
@@ -1,10 +1,44 @@
 import 'package:flutter/material.dart';
 
+import 'dialogs/logout_dialog.dart';
+import 'dialogs/logs_dialog.dart';
+
 class SettingsScreen extends StatelessWidget {
   const SettingsScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return const Placeholder();
+    return LayoutBuilder(
+      builder: (context, constraints) {
+        final screenHeight = constraints.maxHeight;
+
+        return SingleChildScrollView(
+          child: ConstrainedBox(
+            constraints: BoxConstraints(minHeight: screenHeight),
+            child: Column(
+              children: [
+                Card(
+                  child: Column(
+                    mainAxisSize: MainAxisSize.min,
+                    children: [
+                      ListTile(
+                        title: const Text('Logs'),
+                        trailing: const Icon(Icons.open_in_new),
+                        onTap: () => showLogsDialog(context),
+                      ),
+                      ListTile(
+                        title: const Text('Logout'),
+                        trailing: const Icon(Icons.logout),
+                        onTap: () => showLogoutDialog(context),
+                      ),
+                    ],
+                  ),
+                ),
+              ],
+            ),
+          ),
+        );
+      },
+    );
   }
 }