+++ app/CHANGELOG.md
@@ -6,6 +6,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.6] - 18/07/26
+
+## Fixed
+- Inverter.lastSeenAt should be nullable
+- EnergyCard has inconsistend value text size.
+- EnergyCard has a minimum width that can allow clipping of the title text.
+
+## Added
+- Log viewer.
+A dialog can now be opened from the settings screen with the ability to view and copy recent logs.
+
 ## [0.1.5] - 17/07/26 
 
 ### Fixed
@@ -50,4 +61,4 @@ Initial beta app release
 [0.1.3]: https://github.com/ConcenTech/grobase/compare/v0.1.2...v0.1.3
 [0.1.4]: https://github.com/ConcenTech/grobase/compare/v0.1.3...v0.1.4
 [0.1.5]: https://github.com/ConcenTech/grobase/compare/v0.1.4...v0.1.5
-
+[0.1.6]: https://github.com/ConcenTech/grobase/compare/v0.1.5...v0.1.6